Adicionando as classes Service e RestController para Pedido, além da adição da camada DTO para a resolução de inconsistências na transferência de dados através das requisições HTTPS

+++ java/seboDoJoao/src/main/java/br/com/interdisciplinar/sebodojoao/repository/GeneroRepository.java
@@ -3,6 +3,5 @@ package br.com.interdisciplinar.sebodojoao.repository;
 import br.com.interdisciplinar.sebodojoao.model.Genero;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GeneroRepository extends JpaRepository<Genero, Integer> {
-
+public interface GeneroRepository extends JpaRepository<Genero, Long> {
 }
