+++ java/seboDoJoao/src/main/java/br/com/interdisciplinar/sebodojoao/repository/PedidoRepository.java
@@ -3,6 +3,5 @@ package br.com.interdisciplinar.sebodojoao.repository;
 import br.com.interdisciplinar.sebodojoao.model.Pedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
-
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
 }
